refactor: 알림과 FCM 토큰 도메인 분리 #39
